Refactor Strings::Inflection module functions to a global instance
